Fix multiple definitions
